Initial commit. Added the following changes to franksmicro's version: Loopback Mode SID Message Filtering OBD Message Filtering Receive Can Message that uses Read RX Buffer SPI command and checks both buffers. Query OBD that can return 4 bytes of data
Moved all defines to header file for ease of reading.

Added comments to several functions who's code was otherwise untouched. Done in order to better understand how they worked.

Some formatting changes were made because the modifications to this source were never intended to be posted. Decided to commit the changes when I needed to implement filtering and thought that it would be useful to others.